+++ packages/snap-client/src/Client/transforms/searchResponse.ts
@@ -199,7 +199,7 @@ transformSearchResponse.results = (response: searchResponseType) => {
 	return { results: results.map(transformSearchResponse.result) };
 };
 
-transformSearchResponse.result = (rawResult: rawResult, idx: number): SearchResponseModelResult => {
+transformSearchResponse.result = (rawResult: rawResult): SearchResponseModelResult => {
 	const coreFieldValues: SearchResponseModelResultCoreMappings = CORE_FIELDS.reduce((coreFields, key) => {
 		if (typeof rawResult[key as keyof rawResult] != 'undefined') {
 			return {
@@ -240,7 +240,6 @@ transformSearchResponse.result = (rawResult: rawResult, idx: number): SearchResp
 
 	return new Result({
 		id: rawResult.uid,
-		position: idx + 1,
 		mappings: {
 			core: coreFieldValues,
 		},